[Refactor]: createSpace 이미지파일 s3 업로드 로직 퍼사드로 이동.
