+++ src/main/java/com/tenten/linkhub/domain/space/controller/mapper/SpaceApiMapper.java
@@ -2,6 +2,7 @@ package com.tenten.linkhub.domain.space.controller.mapper;
 
 import com.tenten.linkhub.domain.space.controller.dto.space.SpaceCreateApiRequest;
 import com.tenten.linkhub.domain.space.controller.dto.space.SpacesFindByQueryApiRequest;
+import com.tenten.linkhub.domain.space.facade.dto.SpaceCreateFacadeRequest;
 import com.tenten.linkhub.domain.space.service.dto.space.SpaceCreateRequest;
 import com.tenten.linkhub.domain.space.service.dto.space.SpacesFindByQueryRequest;
 import org.mapstruct.InjectionStrategy;
@@ -17,5 +18,7 @@ public interface SpaceApiMapper {
     SpacesFindByQueryRequest toSpacesFindByQueryRequest(SpacesFindByQueryApiRequest request, Pageable pageable);
 
     SpaceCreateRequest toSpaceCreateRequest(SpaceCreateApiRequest request, MultipartFile file);
+
+    SpaceCreateFacadeRequest toSpaceCreateFacadeRequest(SpaceCreateApiRequest request, MultipartFile file);
 
 }
