+++ src/main/java/com/tenten/linkhub/domain/space/facade/mapper/SpaceFacadeMapper.java
@@ -1,0 +1,15 @@
+package com.tenten.linkhub.domain.space.facade.mapper;
+
+import com.tenten.linkhub.domain.space.facade.dto.SpaceCreateFacadeRequest;
+import com.tenten.linkhub.domain.space.service.dto.space.SpaceCreateRequest;
+import com.tenten.linkhub.global.aws.dto.ImageInfo;
+import org.mapstruct.InjectionStrategy;
+import org.mapstruct.Mapper;
+
+@Mapper(componentModel = "spring",
+        injectionStrategy = InjectionStrategy.CONSTRUCTOR
+)
+public interface SpaceFacadeMapper {
+
+    SpaceCreateRequest toSpaceCreateRequest(SpaceCreateFacadeRequest request, ImageInfo imageInfo);
+}
