+++ src/main/java/com/tenten/linkhub/domain/space/service/DefaultSpaceService.java
@@ -8,29 +8,20 @@ import com.tenten.linkhub.domain.space.service.dto.space.SpacesFindByQueryReques
 import com.tenten.linkhub.domain.space.service.dto.space.SpacesFindByQueryResponses;
 import com.tenten.linkhub.domain.space.service.mapper.SpaceMapper;
 
-import com.tenten.linkhub.global.aws.dto.ImageInfo;
-import com.tenten.linkhub.global.aws.dto.ImageSaveRequest;
-import com.tenten.linkhub.global.aws.s3.S3Uploader;
-
 import org.springframework.data.domain.Slice;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import org.springframework.web.multipart.MultipartFile;
 
 import static com.tenten.linkhub.domain.space.model.space.Role.OWNER;
 
 @Service
 public class DefaultSpaceService implements SpaceService{
-    private static final String SPACE_IMAGE_FOLDER = "space-image/";
-    private static final String SPACE_DEFAULT_IMAGE_PATH = "https://team-10-bucket.s3.ap-northeast-2.amazonaws.com/%08space-image/space-default.png";
 
     private final SpaceRepository spaceRepository;
-    private final S3Uploader s3Uploader;
     private final SpaceMapper mapper;
 
-    public DefaultSpaceService(SpaceRepository spaceRepository, S3Uploader s3Uploader, SpaceMapper mapper) {
+    public DefaultSpaceService(SpaceRepository spaceRepository, SpaceMapper mapper) {
         this.spaceRepository = spaceRepository;
-        this.s3Uploader = s3Uploader;
         this.mapper = mapper;
     }
 
@@ -51,10 +42,8 @@ public class DefaultSpaceService implements SpaceService{
                 mapper.toSpaceMember(request, OWNER)
         );
 
-        ImageInfo imageInfo = getImageInfo(request.file());
-
         space.addSpaceImage(
-                mapper.toSpaceImage(imageInfo)
+                mapper.toSpaceImage(request.imageInfo())
         );
 
         return spaceRepository.save(space).getId();
@@ -66,15 +55,6 @@ public class DefaultSpaceService implements SpaceService{
         Space space = spaceRepository.getSpaceJoinSpaceMemberById(spaceId);
 
         return SpaceWithSpaceImageAndSpaceMemberInfo.from(space);
-    }
-
-    private ImageInfo getImageInfo(MultipartFile file) {
-        if (file == null){
-            return ImageInfo.of(SPACE_DEFAULT_IMAGE_PATH, "default-image");
-        }
-
-        ImageSaveRequest imageSaveRequest = ImageSaveRequest.of(file, SPACE_IMAGE_FOLDER);
-        return s3Uploader.saveImage(imageSaveRequest);
     }
 
 }
