+++ src/test/java/com/tenten/linkhub/domain/space/service/DefaultSpaceServiceTest.java
@@ -6,29 +6,22 @@ import com.tenten.linkhub.domain.space.model.space.Space;
 import com.tenten.linkhub.domain.space.model.space.SpaceImage;
 import com.tenten.linkhub.domain.space.model.space.SpaceMember;
 import com.tenten.linkhub.domain.space.repository.space.SpaceJpaRepository;
-import com.tenten.linkhub.domain.space.service.dto.space.SpaceCreateRequest;
 import com.tenten.linkhub.domain.space.service.dto.space.SpacesFindByQueryRequest;
 import com.tenten.linkhub.domain.space.service.dto.space.SpacesFindByQueryResponse;
 import com.tenten.linkhub.domain.space.service.dto.space.SpacesFindByQueryResponses;
-import com.tenten.linkhub.global.aws.dto.ImageInfo;
-import com.tenten.linkhub.global.aws.s3.S3Uploader;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.mockito.BDDMockito;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Sort;
-import org.springframework.mock.web.MockMultipartFile;
 import org.springframework.test.context.TestPropertySource;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.mockito.ArgumentMatchers.any;
 
 @Transactional
 @TestPropertySource(locations = "classpath:/application-test.yml")
@@ -40,9 +33,6 @@ class DefaultSpaceServiceTest {
 
     @Autowired
     private SpaceJpaRepository spaceJpaRepository;
-
-    @MockBean
-    private S3Uploader s3Uploader;
 
     @BeforeEach
     void setUp() {
@@ -72,40 +62,6 @@ class DefaultSpaceServiceTest {
         assertThat(content.get(0).description()).isEqualTo("첫번째 스페이스 소개글");
         assertThat(content.get(0).category()).isEqualTo(Category.KNOWLEDGE_ISSUE_CAREER);
         assertThat(content.get(0).spaceImagePath()).isEqualTo("https://testimage1");
-    }
-
-    @Test
-    @DisplayName("유저는 스페이스를 생성할 수 있다.")
-    void createSpace(){
-        //given
-        MockMultipartFile requestFile = new MockMultipartFile("테스트 이미지3", (byte[]) null);
-        ImageInfo imageInfo = ImageInfo.of("https://testimage3", requestFile.getName());
-        BDDMockito.given(s3Uploader.saveImage(any())).willReturn(imageInfo);
-
-        SpaceCreateRequest spaceCreateRequest = new SpaceCreateRequest(
-                "테스트용 스페이스 이름",
-                "테스트용 스페이스 소개글",
-                Category.ENTER_ART,
-                true,
-                true,
-                true,
-                true,
-                3L,
-                requestFile
-        );
-
-        //when
-        Long savedSpaceId = spaceService.createSpace(spaceCreateRequest);
-
-        //then
-        Space savedSpace = spaceJpaRepository.findById(savedSpaceId).get();
-        SpaceImage spaceImage = savedSpace.getSpaceImages().get(0);
-
-        assertThat(savedSpace.getSpaceName()).isEqualTo("테스트용 스페이스 이름");
-        assertThat(savedSpace.getDescription()).isEqualTo("테스트용 스페이스 소개글");
-        assertThat(savedSpace.getCategory()).isEqualTo(Category.ENTER_ART);
-        assertThat(spaceImage.getPath()).isEqualTo("https://testimage3");
-        assertThat(spaceImage.getName()).isEqualTo("테스트 이미지3");
     }
 
     private void setupData() {
